moved flashinfer functions and updated for compressed_tensors_w4a4_nvfp4.py
