+++ vllm/_custom_ops.py
@@ -10,7 +10,6 @@ import vllm.envs as envs
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
 from vllm.scalar_type import ScalarType
-from vllm.utils.flashinfer import has_flashinfer
 
 logger = init_logger(__name__)
 
@@ -642,48 +641,6 @@ if hasattr(torch.ops._C, "ggml_moe_a8_vec"):
                            device=W.device)
 
 
-if has_flashinfer():
-    from flashinfer import mm_fp4 as flashinfer_mm_fp4_
-
-    @torch.library.custom_op(
-        "vllm::flashinfer_mm_fp4",
-        mutates_args=[],
-        device_types="cuda",
-    )
-    def flashinfer_mm_fp4(
-        A: torch.Tensor,
-        B: torch.Tensor,
-        A_scale: torch.Tensor,
-        B_scale: torch.Tensor,
-        g_scale: torch.Tensor,
-        dtype: torch.dtype,
-        backend: str,
-    ) -> torch.Tensor:
-        return flashinfer_mm_fp4_(A,
-                                  B,
-                                  A_scale,
-                                  B_scale,
-                                  g_scale,
-                                  dtype,
-                                  block_size=16,
-                                  backend=backend)
-
-    @torch.library.register_fake("vllm::flashinfer_mm_fp4", )
-    def flashinfer_mm_fp4_fake(
-        A: torch.Tensor,
-        B: torch.Tensor,
-        A_scale: torch.Tensor,
-        B_scale: torch.Tensor,
-        g_scale: torch.Tensor,
-        dtype: torch.dtype,
-        backend: str,
-    ) -> torch.Tensor:
-        return torch.empty(A.shape[0],
-                           B.shape[1],
-                           dtype=dtype,
-                           device=A.device)
-
-
 # cutlass
 def cutlass_scaled_mm_supports_fp4(cuda_device_capability: int) -> bool:
     return torch.ops._C.cutlass_scaled_mm_supports_fp4(cuda_device_capability)
@@ -713,33 +670,6 @@ def cutlass_scaled_fp4_mm(a: torch.Tensor, b: torch.Tensor,
     torch.ops._C.cutlass_scaled_fp4_mm(out, a, b, block_scale_a, block_scale_b,
                                        alpha)
     return out
-
-
-def flashinfer_scaled_fp4_mm(a: torch.Tensor, b: torch.Tensor,
-                             block_scale_a: torch.Tensor,
-                             block_scale_b: torch.Tensor, alpha: torch.Tensor,
-                             out_dtype: torch.dtype,
-                             backend: str) -> torch.Tensor:
-    assert a.ndim == 2 and b.ndim == 2
-    assert block_scale_a.ndim == 2 and block_scale_b.ndim == 2
-    assert a.stride(-1) == 1 and b.stride(-1) == 1
-    assert a.shape[1] == b.shape[1]
-    assert block_scale_a.shape[1] == a.shape[1] // 8
-    assert block_scale_b.shape[1] == b.shape[1] // 8
-
-    if backend == "cutlass":
-        block_scale_a = block_scale_a.view(torch.uint8)
-        block_scale_b = block_scale_b.view(torch.uint8)
-
-    return flashinfer_mm_fp4(
-        a,
-        b.t(),
-        block_scale_a,
-        block_scale_b.t(),
-        alpha,
-        out_dtype,
-        backend=backend,
-    )
 
 
 def cutlass_scaled_mm_supports_fp8(cuda_device_capability: int) -> bool:
